Stagedef endian fixing progress
I made a realization though: this isn't going to work on 64-bit
platforms because the stagedef structs will have different sizes because
the pointer sizes will be different. I think there will need to be two
sets of stagedef structs, one with exact sizes meant to load the smb2
stagedef directly into, and another set that it is converted to that is
used as the _actual_ stagedef.